Feature: Add geoCodeLoop function to call geoCode10Addresses at 10 second intervals.
The Google Maps API geocaching service allows for 10 initial requests which are
then refreshed at 1 request per second to a maximum of the original 10 requests.
The geoCodeLoop function is called taking geoCodeArray (an array of IDs), geocoder
(the Google Maps geocoder object), data (the data of the ajax request), and
geoCodePassesLeft (the length of geoCodeArray which is decrimented by 10 each time
geoCode10Addresses is called and 10 murals are updated). It checks to see if there
are still murals to geocode, and if there are, calls geoCode10Addresses, decriments
geoCodePassesLeft by 10, then after a 10 second delay recursively calls geoCodeLoop
again. This is done until geoCodePassesLeft > 0 is no longer true and all of the
murals have been geocoded. Once that happens, it simply sends a message to the
console that the geocoding is done and does not recursively call itself again.