Refactor: Combined geoCode10Addresses and geoCodeLoop into one function
This major refactoring of geocoding functions fixes the issue of only geocoding
the first ten addresses. geoCodePasses left is now geoCodeIndex which also
keeps track of which index in geoCodeArray is being geocoded. The conditional
of whether to recursively call the function again has been moved to the end
of the function. This conditional has also been modified to handle the case
when there are less than 10 murals left to geocode. In this case, it passes
the number of iterations (as 'iterations') to the function, which is 10 by
default.